+++ src/main/java/dominio/Core/Obtener_Zapatos.java
@@ -1,8 +1,8 @@
-package dominio.Core;
+package main.java.dominio.Core;
 
 import java.util.ArrayList;
 
-import dominio.GestorZapatos;
+import main.java.dominio.GestorZapatos;
 
 public class Obtener_Zapatos {
