+++ src/main/java/dominio/Core/Obtener_Zapatos_Model.java
@@ -1,32 +1,33 @@
-package dominio.Core;
+package main.java.dominio.Core;
 
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
-import dominio.GestorZapatos;
-import dominio.Enums.Color;
+import main.java.dominio.GestorZapatos;
 
-public class Obtener_Zapatos_talle {
+public class Obtener_Zapatos_Model {
 
     private GestorZapatos miGestorZapatos;
 
-    public Obtener_Zapatos_talle(GestorZapatos gz){
+    private Obtener_Zapatos_Model(GestorZapatos gz){
         this.miGestorZapatos = gz;
     }
-    // talle por empate modelo
-		public ArrayList<Zapato> getZapatos(Color miColor) {
+
+    //por modelo y ante empate por color 
+		public ArrayList<Zapato> getZapatos(String modelo) {
+			
 			ArrayList<Zapato> aux = miGestorZapatos.getZapatos();
 			ArrayList<Zapato> aux2=aux.stream()
-						.filter(p -> p.getMiColor().compareTo(miColor)==0)
+						.filter(p -> p.getModelo().compareToIgnoreCase(modelo)==0)
 						.sorted(new Comparator<Zapato>() {
 							@Override
 							public int compare(Zapato z1, Zapato z2) {
-								if (z1.getTalle()-z2.getTalle()==0) {
-									if (z1.getModelo().compareTo(z2.getModelo())==0) {
+								if (z1.getTalle()-z2.getTalle() == 0) {
+									if (z1.getMiColor().compareTo(z2.getMiColor())==0) {
 										return z1.getCodigo().compareTo(z2.getCodigo());
 									}
-									return z1.getModelo().compareTo(z2.getModelo())*-1;
+									return z1.getMiColor().compareTo(z2.getMiColor())*-1;
 								}
 								return z1.getTalle()-z2.getTalle();
 							}
@@ -35,6 +36,5 @@ public class Obtener_Zapatos_talle {
 			
 			return aux2;
 		}
-
 
 }
