+++ src/main/java/dominio/Core/Obtener_Zapatos_talle.java
@@ -1,33 +1,32 @@
-package dominio.Core;
+package main.java.dominio.Core;
 
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.stream.Collectors;
 
-import dominio.GestorZapatos;
+import main.java.dominio.GestorZapatos;
+import main.java.dominio.Enums.*;
 
-public class Obtener_Zapatos_Model {
+public class Obtener_Zapatos_talle {
 
     private GestorZapatos miGestorZapatos;
 
-    private Obtener_Zapatos_Model(GestorZapatos gz){
+    public Obtener_Zapatos_talle(GestorZapatos gz){
         this.miGestorZapatos = gz;
     }
-
-    //por modelo y ante empate por color 
-		public ArrayList<Zapato> getZapatos(String modelo) {
-			
+    // talle por empate modelo
+		public ArrayList<Zapato> getZapatos(Color miColor) {
 			ArrayList<Zapato> aux = miGestorZapatos.getZapatos();
 			ArrayList<Zapato> aux2=aux.stream()
-						.filter(p -> p.getModelo().compareToIgnoreCase(modelo)==0)
+						.filter(p -> p.getMiColor().compareTo(miColor)==0)
 						.sorted(new Comparator<Zapato>() {
 							@Override
 							public int compare(Zapato z1, Zapato z2) {
-								if (z1.getTalle()-z2.getTalle() == 0) {
-									if (z1.getMiColor().compareTo(z2.getMiColor())==0) {
+								if (z1.getTalle()-z2.getTalle()==0) {
+									if (z1.getModelo().compareTo(z2.getModelo())==0) {
 										return z1.getCodigo().compareTo(z2.getCodigo());
 									}
-									return z1.getMiColor().compareTo(z2.getMiColor())*-1;
+									return z1.getModelo().compareTo(z2.getModelo())*-1;
 								}
 								return z1.getTalle()-z2.getTalle();
 							}
@@ -36,5 +35,6 @@ public class Obtener_Zapatos_Model {
 			
 			return aux2;
 		}
+
 
 }
