+++ src/main/java/dominio/Enums/Color.java
@@ -1,4 +1,4 @@
-package dominio.Enums;
+package main.java.dominio.Enums;
 
 public enum Color {
 	ROJO,NEGRO,AZUL,BLANCO;
