+++ src/main/java/dominio/Enums/Pais.java
@@ -1,4 +1,4 @@
-package dominio.Enums;
+package main.java.dominio.Enums;
 
 public enum Pais {
     ITALIA,ARGENTINA,MACEDONIA,URUGUAY;
