+++ src/main/java/dominio/Exceptions/ExceptionZapatoModeloNull.java
